override toString and equals
all the classes are belong to Object, so you can use everyting in Object,including toString and equals. toSting in object outputs name and positon. equals in object outputs if they point to the same object or not. so you can wirte your own tostring and equals to override them. override tostring can output the content in string. equals can compare the content of two 对象变量